rag-chatbot/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
requests
pydantic
httpx